Prepare to add in feature where we open on demand
This is gonna make it slow, so we turn it off by default. Furthermore,
if this is turned on and there are race conditions.. it will fail with
LevelDBError IOError

So I think the ultimate goal is to write a leveldb-server with 2i
capability (maybe cache..) as a super lightweight (no cache)
replacement for Riak... Shall see.